feat: proxy @waku/core and @waku/utils through @waku/sdk
